requirements.txt
fastapi[all]==0.116.1
PyJWT==2.10.1
httpx==0.28.1
sqlalchemy==2.0.42
opencv-python==4.12.0.88
requests==2.32.4